Upgrading router version (#19)
---



**Size M:** This PR changes include 76 lines and should take
approximately 30-60 minutes to review



---



 **DeputyDev generated PR summary:** 

The provided pull request (PR) updates the `package.json` file of a
project named "catalyst-core". Specifically, it makes the following
changes:

1. **Version Update**: 
- The version of the "catalyst-core" package is updated from
`0.0.1-beta.8` to `0.0.1-beta.9`.

2. **Dependency Update**:
- The version of the `@tata1mg/router` dependency is updated from
`0.0.1-beta.5` to `0.0.1-beta.6`.

These changes suggest that there might have been some enhancements, bug
fixes, or other updates in the `@tata1mg/router` package that the
project wants to incorporate. The version bump of the "catalyst-core"
package itself indicates that the project is likely preparing for a new
release, possibly to include these updates.

---------

Co-authored-by: Mayan Sharma <mayansharma@C02HFD2GQ6L4.local>