Updates changelog (#49)
---



 **DeputyDev generated PR summary:** 



---



**Size XS:** This PR changes include 4 lines and should take
approximately 5-15 minutes to review



---

The pull request updates the `changelog.md` file by adding a new entry
for version `0.0.1-beta.13` dated `04-04-2025`. This entry documents a
bug fix related to the client fetcher when importing components through
a tool or library called "loadable". The changelog is a file that keeps
a chronological list of changes made to the project, which includes new
features, bug fixes, and other updates.

---

DeputyDev generated PR summary until
7bf4c784db1c95e3e6ee9ff9bf1269c3429c4776

Co-authored-by: Mayan Sharma <mayansharma@G206F1WHVF.local>