Add support for native keyboard and webview resizing in Android (#99)
When the user focuses on any input field in the webview, the keyboard
would open using the browser API and would overlap the content of the
webview blocking the input field. This PR opens the native keyboard when
user focusses on any input field and resizes the webview with a scroll.

---------

Co-authored-by: utkarsh-1mg <utkarsh.srivastava@1mg.com>